+++ tests/files/src/0764_closure_bind.php
@@ -1,0 +1,17 @@
+<?php
+
+class C764 {
+    private static function bindAndCheck()
+    {
+        $byFn = function (stdClass $x) {
+            return isset($x->prop);
+        };
+
+        $boundByFn = Closure::bind($byFn, null, self::class);
+        if ($boundByFn === false) {  // Not PhanImpossibleTypeComparison
+            throw new RuntimeException();
+        }
+
+        return $boundByFn;
+    }
+}